+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -9,6 +9,12 @@ however, it has to be formatted properly to pass verification tests.
 
 ## [Unreleased]
 
+### Fixed
+
+#### Actions
+
+- Fixed `InputActionTrace.ActionEventPtr.ReadValueAsObject` leading to `InvalidCastException` when trying to read values that came from composite bindings.
+
 ### Added
 
 #### Actions