+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 Due to package verification, the latest version below is the unpublished version and the date is meaningless.
 however, it has to be formatted properly to pass verification tests.
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed Action with multiple bindings becoming unresponsive after a Hold interaction was performed. ([case 1239551](https://issuetracker.unity3d.com/issues/input-system-hold-interaction-makes-an-input-action-unresponsive-when-2-or-more-binding-are-attached-to-the-same-input-action))
+
 ## [1.1.0-preview.1] - 2020-08-20
 
 >__The minimum version requirement for the Input System package has been moved up to 2019.4 LTS.__
@@ -39,7 +45,6 @@ however, it has to be formatted properly to pass verification tests.
 - Fixed memory corruption in `InputEventTrace.AllocateEvent` ([case 1262496](https://issuetracker.unity3d.com/issues/input-system-crash-with-various-stack-traces-when-using-inputactiontrace-dot-subscribetoall))
   * Manifested itself, for example, as crashes when using `InputActionTrace.SubscribeToAll`.
 - AxisControls and Vector2Controls' X and Y subcontrols on XR devices now have a minimum range of -1 and a maximum range of 1. This means they can now properly respond to modifiers and interactions in the binding system.
-- Fixed Action with multiple bindings becoming unresponsive after a Hold interaction was performed. ([case 1239551](https://issuetracker.unity3d.com/issues/input-system-hold-interaction-makes-an-input-action-unresponsive-when-2-or-more-binding-are-attached-to-the-same-input-action))
 
 #### Actions