+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -9,6 +9,14 @@ however, it has to be formatted properly to pass verification tests.
 
 ## [Unreleased]
 
+### Fixed
+
+#### Actions
+
+- Fixed not being able to stack a `MultiTap` on top of a `Tap` ([case 1261462](https://issuetracker.unity3d.com/issues/multi-tap-and-tap-interactions-in-the-same-action-doesnt-work-properly)).
+
+## [1.1.0-preview.1] - 2020-08-20
+
 >__The minimum version requirement for the Input System package has been moved up to 2019.4 LTS.__
 
 ### Changed